UPDATE HECHO POR EDER

+++ trunk/ITradeMovil/src-gen/com/itrade/model/Producto.java
@@ -7,8 +7,7 @@ package com.itrade.model;
 public class Producto {
 
     private Long id;
-    private Long IdProducto;
-    /** Not-null value. */
+    private int IdProducto;
     private String Descripcion;
     private Double Precio;
     private Integer Stock;
@@ -23,7 +22,7 @@ public class Producto {
         this.id = id;
     }
 
-    public Producto(Long id, Long IdProducto, String Descripcion, Double Precio, Integer Stock, String Activo, Integer IdCategoria, Integer IdMarca) {
+    public Producto(Long id, int IdProducto, String Descripcion, Double Precio, Integer Stock, String Activo, Integer IdCategoria, Integer IdMarca) {
         this.id = id;
         this.IdProducto = IdProducto;
         this.Descripcion = Descripcion;
@@ -42,20 +41,18 @@ public class Producto {
         this.id = id;
     }
 
-    public Long getIdProducto() {
+    public int getIdProducto() {
         return IdProducto;
     }
 
-    public void setIdProducto(Long IdProducto) {
+    public void setIdProducto(int IdProducto) {
         this.IdProducto = IdProducto;
     }
 
-    /** Not-null value. */
     public String getDescripcion() {
         return Descripcion;
     }
 
-    /** Not-null value; ensure this value is available before it is saved to the database. */
     public void setDescripcion(String Descripcion) {
         this.Descripcion = Descripcion;
     }
