+++ trunk/ITradeMovil/src-gen/com/itrade/model/Prospecto.java
@@ -7,9 +7,8 @@ package com.itrade.model;
 public class Prospecto {
 
     private Long id;
-    private Long IdProspecto;
-    private Integer IdPersona;
-    /** Not-null value. */
+    private Integer IdProspecto;
+    private int IdPersona;
     private String Razon_Social;
     private String RUC;
     private Double Latitud;
@@ -26,7 +25,7 @@ public class Prospecto {
         this.id = id;
     }
 
-    public Prospecto(Long id, Long IdProspecto, Integer IdPersona, String Razon_Social, String RUC, Double Latitud, Double Longitud, String Direccion, Integer IdCobrador, Integer IdUsuario, String Activo) {
+    public Prospecto(Long id, Integer IdProspecto, int IdPersona, String Razon_Social, String RUC, Double Latitud, Double Longitud, String Direccion, Integer IdCobrador, Integer IdUsuario, String Activo) {
         this.id = id;
         this.IdProspecto = IdProspecto;
         this.IdPersona = IdPersona;
@@ -48,28 +47,26 @@ public class Prospecto {
         this.id = id;
     }
 
-    public Long getIdProspecto() {
+    public Integer getIdProspecto() {
         return IdProspecto;
     }
 
-    public void setIdProspecto(Long IdProspecto) {
+    public void setIdProspecto(Integer IdProspecto) {
         this.IdProspecto = IdProspecto;
     }
 
-    public Integer getIdPersona() {
+    public int getIdPersona() {
         return IdPersona;
     }
 
-    public void setIdPersona(Integer IdPersona) {
+    public void setIdPersona(int IdPersona) {
         this.IdPersona = IdPersona;
     }
 
-    /** Not-null value. */
     public String getRazon_Social() {
         return Razon_Social;
     }
 
-    /** Not-null value; ensure this value is available before it is saved to the database. */
     public void setRazon_Social(String Razon_Social) {
         this.Razon_Social = Razon_Social;
     }
